Move cluster_name part into the end of global transaction identifier.
Previously gid that the commands for prepared transactions use consisted
of cluster_name, full transaction ID and OID of user mapping with the format
"pgfdw_<cluster_name>_<fxid>_<umid>". But if cluster_name includes
the underscore character, the logic based on the format of gid could not
work well.

To handle the case, this commit updates the format of gid to
"pgfdw_<fxid>_<umid>_<cluster_name>".