Don't use two phase commit protocol if not necessary.
If only read transaction ran or only a single server executed
write transaction, two phase commit protocol is not required
to commit the local and foreign transactions. In this case
we can safely issue COMMIT to the foreign servers and complete
the local transaction. Which would increase the performance
on COMMIT operation by skipping two phase commit.

This commit adds the mechanism to track whether each foreign
server has executed a write transaction, i.e., modified the foreign
tables. Then if write transactions were performed on two or more
servers, we determine that two phase commit protocol is required
to commit the transactions, otherwise not.

Note that a foreign transaction is marked as modified if
it inserts/updates/deletes/truncates the foreign tables. Even if
the scan on foreign tables triggers write operations on the foreign
server, its foreign transaction is NOT marked as modified.