Make pg_resolve_foreign_prepared_xacts_all() keep working even when SET ROLE fails.
If non-superuser executes pg_resolve_foreign_prepared_xacts_all() and there are
user mappings for superusers, SET ROLE command inside the function would fail
because it tries to set that non-superuser to superuser of the user mapping.
Previously if this failure happens, the function reported an error and exited.
But it was inconvenient because non-superuser always fails to call the function
if there are user mappings for superuser.

This commit changes the function so that it reports notice message and
keeps processing instead of exiting with an error, when SET ROLE fails.