Use dblink_connect() and dblink_exec() in pg_resolve_foreign_prepared_xacts().
Previously pg_resolve_foreign_prepared_xacts() used dblink() to execute
a transaction command on the foreign server. But it required that
new connection needs to be established for every time when a command
is executed. Which could decrease the performance of the function.

This commit changes pg_resolve_foreign_prepared_xacts() so that it uses
dblink_connect() and dblink_exec() instead. The function can establish
new connection to the foreign server only once and reuse that connection
to execute the transaction command. This would decrease the number
of connection establishments and improve the performance of this function
especially when there are lots of foreign prepared transactions to resolve.