Improve SQL accessing to pg_locks to check whether the transaction is running.
Previously we checked whether the record with locktype 'transactionid' and
the given PID, XID was found in pg_locks, and if found we determined that
the target transaction was running. But these filtering conditions were not
enough and the record could be found even when the target transaction
was not in progress.

To avoid such false detection, this commit adds the filtering conditions of
mode=ExclusiveLock and granted=true into the SQL.