Improve message logged when current user cannot connect to remote server.
pg_resolve_foreign_prepared_xacts_all() and
pg_min_fxid_foreign_prepared_xacts_all() set current user to the user of
user mapping in order to connect to remote server. If current user is not
a superuser nor it's not a member of the target user, it's not allowed to be
set to the target one, i.e., it cannot connect to the remote server.
Previously, in this case, log message "could not ..." was reported and
it looked like it indicates serious situation even though it's not.

This commit changes those function so that they skip processing the target
server at all in that case and report "less serious" log message "skipping ...".