Improve handling of public user mapping.
Previously when pg_resolve_foreign_prepared_xacts_all() found public
user mapping, its caller user was used to use that public mapping and
connect to a foreign server. However if there was user mapping for that
user itself on the server, public mapping could not used because of
PostgreSQL specification. This could lead to the issue that any foreign
prepared transactions created via such public mapping could never be
resolved.

This commit improves pg_resolve_foreign_prepared_xacts_all() so that
it finds out proper user to use public mapping. When public user mapping
is found, the user who has right to use the server, whom current user
has membership in, and who has no user mapping for itself to the server
is chosen as a user to connect to a foreign server.