Use superuser if both superuser and non- are picked up for public mapping.
This commit changes pg_user_for_public_mapping() so that it chooses
superuser preferentially if both superuser and non-superuser are picked up
as user who can use public mapping. Because this case happens only when
superuser calls this function and which means that it basically wants to
use superuser for processing as much as possible.