Change prefix of functions from "pg_" to "pgfdw_plus_".
Previously "pg_" was used as a prefix of names of functions that
postgres_fdw_plus supported. But those names could conflict with
the names of functions that PostgreSQL core or its contrib modules
will add in the future. To prevent such conflictions, this commit
changes the prefix of those functions from "pg_" to "pgfdw_plus_".