postgres_fdw_plus: Add support for parallel prepare.
With 2PC enabled, postgres_fdw prepares remote transactions opened on
remote server(s) in a local transaction one by one when the local transaction
commits. This patch allows it to prepare the remote transactions in parallel
to improve performance. This is enabled by the existing postgres_fdw's option
"parallel_commit".