postgres_fdw: Allow postgres_fdw_get_connections() to check connection
If requested, postgres_fdw_get_connections() can verify the status of
connections that are establieshed by postgres_fdw. This check wil be
done by polling the socket. This feature is currently available only
on systems that support the non-standard POLLRDHUP extension to the
poll system call, including Linux.

"closed" column is set to false if existing connection is not closed
by the remote peer.