Add log_autovacuum_analyze_min_duration
The log output functionality of log_autovacuum_min_duration applies to
both VACUUM and ANALYZE, so it is not possible to separate the VACUUM
and ANALYZE log output thresholds. Logs are likely to be output only for
VACUUM and not for ANALYZE.

Therefore, we decided to separate the threshold for log output of
VACUUM by autovacuum (log_autovacuum_vacuum_min_duration) and the
threshold for log output of ANALYZE by autovacuum
(log_autovacuum_analyze_min_duration).