Don't strip $libdir from LOAD command
The 4f7f7b03758 commit implement the extension_control_path GUC, and to
make it work it was decided that we should strip the $libdir/ on
module_pathname from .control files, so that extensions don't need to
worry about this change.

This strip logic was implemented on expand_dynamic_library_name() which
works fine when executing the SQL functions from extensions but this
function is also called when the LOAD command is executed, and since the
user may explicitly pass the $libdir prefix on LOAD parameter we should
not strip in this case.

This commit fix this issue by moving the strip logic from
expand_dynamic_library_name() to load_external_function() that is called
when the running the SQL script from extensions.